refactor(runtime): move runtime interfaces to core
