+++ packages/graphback-core/src/index.ts
@@ -1,6 +1,5 @@
 
 //Top level API
-
 export * from './crud'
 export * from './plugin/GraphbackPlugin'
 export * from './plugin/getModelTypesFromSchema'
@@ -13,4 +12,5 @@ export * from './utils/metadataAnnotations';
 export * from './plugin/GraphbackCoreMetadata'
 export * from './relationships/RelationshipMetadataBuilder';
 export * from './relationships/relationshipHelpers';
+export * from './runtime'
 export * from './db';
