--- packages/graphback-runtime/src/GraphbackOrderBy.ts
@@ -1,9 +1,0 @@
-type SortDirection = 'asc' | 'desc'
-
-/**
- * @todo move this into a common file
- */
-export interface GraphbackOrderBy {
-  order?: SortDirection
-  field: string
-}
