--- packages/graphback-runtime/src/GraphbackPage.ts
@@ -1,7 +1,0 @@
-/**
- * Interface represents GraphbackPage type
- */
-export interface GraphbackPage {
-    limit?: number,
-    offset?: number,
-};
