+++ packages/graphback/src/index.ts
@@ -1,5 +1,4 @@
 export * from '@graphback/core'
-export * from '@graphback/runtime'
 export * from './buildGraphbackAPI'
 export * from './GraphbackConfig'
 export * from './GraphbackGenerator'
